feature(navigation) add BottomBar and RootNav for improved navigation structure, implement navigation graphs for screens

+++ PokemonApp/app/src/main/java/com/netexlearning/pokemon/ui/pokemonlist/PokemonListScreen.kt
@@ -33,7 +33,7 @@ fun PokemonListScreen(
     PokemonListScreen(
         pokemonList = pokemonList,
         modifier = modifier,
-        onItemClick = { pokemonNavigation.navigateToDetail(it.id) },
+        onItemClick = { pokemonNavigation.navigateToDetail(it.id.toInt()) },
         onLoadMore = { viewModel.loadNextPage() }
     )
 }
